refactor(mysql): keep legacy base separate
